cheaper N1 station temporary fix N1 station in Storihjalli, near Skemmuvegur in Reykjavik IRL observation tells us the fuel price there is 8 ISK cheaper than their global price. At some point N1 added the following comment to their global one price webpage:
    (https://www.n1.is/thjonusta/eldsneyti)
    "* algengasta sjalfsafgreidsluverdid"

meaning this one global price might not be the case on all N1 stations, I'd like to see N1 then move from showing one global price to show individual prices for stations or list stations that have different prices than the global one, this is a matter of transparency to N1 customers.

Note: this above price "fix" is bad and will very likely become wrong in near future, however, I believe it to be a good thing to add this for now.