use Selenium Webdriver for Orkan webpage
resolves issue #16

Orkan moved to a client side rendered website via the Blazor JS front-end library. To begin with we were able to receive server side rendered webpage by providing a well known google bot webscraper user-agent string, this however stopped working in 2025-07, we still got server side rendered page, but the price data table we were parsing prices out of was no longer included in the server side rendered HTML.

I made a couple of attempting to do manual Blazor socket interaction to retrieve the required data, but those attempts were unfortunately fruitless, I'm guessing I'd need to dig deeper into the Blazor client engine to figure out how to properly manipulate the socket connection to serve the appropriate price data table elements, which I'm currently not motivated enough to start the process of doing.

I was hesitant to begin running a full browser instance for this, and attempted for a month to get in touch with someone at Orkan to resolve this by either fixing the SSR so it again would include the price data or get an alternative webpage to fetch or even a data API, but unfortunately I never got any replies from Orkan.

So here we are, forced for the time being to settle for running a full Selenium Webdriver Firefox browser instance in headless mode, have it render the Orkan Blazor JS client until the appropriate elements have finished rendering to DOM, then grab the HTML for those elements and read price data from it.